+++ codigo/EEG/app.py
@@ -38,15 +38,15 @@ def main():
             duration = n / fs
             number_of_channels = int(input("Enter number of channels: "))
             while True:
-                signals = reception.get_real_data(channels=number_of_channels, n=n, fs=fs, filter=True)
+                signals = reception.get_real_data(channels=number_of_channels, n=n)
                 for ch, signal in signals.items():
                     t = np.arange(len(signals[ch])) / fs
                     graphingTerminal.graph_signal_voltage_time(t=t, signal=np.array(signal), title="Signal {}".format(ch))
                 
-                t0, alpha0, beta0, gamma0, delta0, theta0 = preprocessing.preprocess_signal(n=n, duration=duration, fs=fs, eeg_data=signal['ch0'])
-                t1, alpha1, beta1, gamma1, delta1, theta1 = preprocessing.preprocess_signal(n=n, duration=duration, fs=fs, eeg_data=signal['ch0'])
-                t2, alpha2, beta2, gamma2, delta2, theta2 = preprocessing.preprocess_signal(n=n, duration=duration, fs=fs, eeg_data=signal['ch0'])
-                t3, alpha3, beta3, gamma3, delta3, theta3 = preprocessing.preprocess_signal(n=n, duration=duration, fs=fs, eeg_data=signal['ch0'])
+                t0, alpha0, beta0, gamma0, delta0, theta0 = preprocessing.preprocess_signal(n=n, duration=duration, fs=fs, eeg_data=np.array(signals['ch0']))
+                t1, alpha1, beta1, gamma1, delta1, theta1 = preprocessing.preprocess_signal(n=n, duration=duration, fs=fs, eeg_data=np.array(signals['ch1']))
+                t2, alpha2, beta2, gamma2, delta2, theta2 = preprocessing.preprocess_signal(n=n, duration=duration, fs=fs, eeg_data=np.array(signals['ch2']))
+                t3, alpha3, beta3, gamma3, delta3, theta3 = preprocessing.preprocess_signal(n=n, duration=duration, fs=fs, eeg_data=np.array(signals['ch3']))
                 
                 print("EEG data processed")
